Pass target option to test-suite
When lnt calls test-suite's configure script, it doesn't pass in a
target option and will get configured for the host instead. If the host
and target are not the same, then variables like TARGET_OS will not be
what's expected, so you won't really be able to compare against them.

So if for example I ran lnt like this from a Linux host:

lnt runtest nt --sandbox ./sandbox --cc i386-pc-solaris2.11-clang
--test-suite ./test-suite --remote --remote-port=22
--remote-host=target_board --remote-user=user

Then test-suite would end up with TARGET_OS=Linux in the Makefile's,
instead of the expected SunOS.

Patch by Michael Daniels